sample-app/app/currency/requirements.txt
pylint==2.4.0
google-cloud==0.34.0
google-cloud-secret-manager==2.0.0
berglas-python==0.3.3
Flask==1.1.2
kafka-python==2.0.1
gunicorn==20.0.4
cloudevents==0.2.4
requests==2.23.0
alpha-vantage==2.2.0